Refactor the tests, rename some constants, split cloud CLI tests into two (registration/querying and metadata)
